Add files via upload
The ArrayBuffer Class
The ArrayBuffer class has a member variable m_buffer which points to a dynamically created array with the size m_capacity. This class acts as a node in a linked list. It has a member variable m_next which points to the next ArrayBuffer object in the linked list. An ArrayBuffer object is a circular array and represents a queue.

The ListBuffer Class
The class ListBuffer has a pointer to an object of type ArrayBuffer. The pointer is stored in the member variable m_cursor. The ListBuffer is a circular singly linked list. The first node of the linked list has a buffer with the size indicated by the member variable m_minBufCapacity. This value is passed through the constructor. The enqueue function inserts data into this buffer. Once the buffer is full, the enqueue function inserts a new node (ArrayBuffer object) into the linked list and starts to insert data in the new buffer. The size of the new buffer is a factor of the size of the buffer in the previous node. We use the constant global variable INCREASE_FACTOR to define the size of the new buffer. For example, if the size of the buffer is 100, and it becomes full, we create a new buffer with the size (INCREASE_FACTOR x 100).

To limit the size of buffer arrays we use the constant global variable MAX_FACTOR. For example, if the minimum size of buffer is 100, the maximum size of a buffer cannot exceed (MAX_FACTOR x 100). As soon as we reach to the maximum size, the next buffer will be created with the minimum size which is defined by the variable member m_minBufCapacity.

To remove data from the queue, the dequeue function removes the data from the start index in the oldest buffer (ArrayBuffer object) of the linked list. If a buffer becomes empty, the dequeue function removes the buffer from the linked list. If there is only one node in the linked list and its buffer is empty, we do not remove the node. Always there is at least one node in the linked list.